assignments/02._serverCom/python/pyproject.toml
[tool.poetry]
name = "python"
version = "0.1.0"
description = ""
authors = ["William Omø <williamjean.k.o@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.20.0"
fastapi = "^0.92.0"
requests = "^2.28.2"
xmltodict = "^0.13.0"
pyyaml = "^6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
